Add removing *.localized files

+++ etc/init/ch_lang_home_dir.sh
@@ -7,21 +7,35 @@ set -u
 echo -n "Translate home directory into English? (y/N) "
 read
 if [[ "$REPLY" =~ ^[Yy]$ ]]; then
-    if [[ $OSTYPE == linux* ]]; then
-        LANG=C xdg-user-dirs-gtk-update
+    if [[ $OSTYPE == darwin* ]]; then
+        rm -vf ~/Desktop/.localized
+        rm -vf ~/Documents/.localized
+        rm -vf ~/Downloads/.localized
+        rm -vf ~/Library/.localized
+        rm -vf ~/Movies/.localized
+        rm -vf ~/Music/.localized
+        rm -vf ~/Pictures/.localized
+        rm -vf /Applications/.localized
 
-    elif [[ $OSTYPE == darwin* ]]; then
+        declare error=0
         sudo -v
-        # CF: http://qiita.com/is0me/items/0b7b846f1f0860629950
-        if [ $(echo "$(sw_vers -productVersion | cut -d . -f1,2) > 10.8" | bc) -eq 1 ]; then
-            cd /System/Library/CoreServices/SystemFolderLocalizations/ja.lproj
+
+        # cf. http://qiita.com/is0me/items/0b7b846f1f0860629950
+        cd /System/Library/CoreServices/SystemFolderLocalizations
+        if [[ -d ja.lproj && -d en.lproj ]]; then
+            cd ja.lproj
+        elif [[ -d Japanese.lproj ]]; then
+            cd Japanese.lproj
         else
-            cd /System/Library/CoreServices/SystemFolderLocalizations/Japanese.lproj
+            error=1
         fi
 
-        mv SystemFolderLocalizations.strings SystemFolderLocalizations.strings.back
-        cp ../en.lproj/SystemFolderLocalizations.strings .
-
-        killall Finder
+        if [[ $error -eq 0 ]]; then
+            sudo mv SystemFolderLocalizations.strings SystemFolderLocalizations.strings.back
+            sudo cp ../en.lproj/SystemFolderLocalizations.strings .
+            killall Finder
+        fi
     fi
+else
+    LANG=C xdg-user-dirs-gtk-update
 fi
